Refactor 'NumKeypad' module to match tab name 'Record' for consistency with other pages.
